add manage listings

+++ app/Http/Controllers/ListingController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Database\Eloquent\Collection;
 use PhpParser\Node\Expr\List_;
+use Illuminate\Routing\Controller;
 
 class ListingController extends Controller
 {
@@ -58,6 +60,11 @@ class ListingController extends Controller
 
     // Update listing data
     public function update(Listing $listing){
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = request()->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -79,7 +86,17 @@ class ListingController extends Controller
 
     // Delete listing 
     public function destroy(Listing $listing){
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
-        return redirect('/')->with('message', 'Listing deleted successfully!');
+        return redirect('/listings/manage')->with('message', 'Listing deleted successfully!');
+    }
+
+    // Manage listings 
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings]);
     }
 }
